mcp_server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
neo4j==5.14.1
pydantic==2.4.2 